Improvements to Edit Widget operator
- De-duplicate ensure_widget() function
- Editing multiple bone shapes at once is no longer supported.
This functionality was never used, and it caused a conflict as well:
A bone shape that already had a widget assigned couldn't be
assigned a different one.
This still can't be done without selecting multiple bones,
so this can still be improved.
Will probably need to be split into two operators.

Added a pie menu idea to my work_todo.txt.