Basic structure of properties
This includes a clean way for rig types to register their
parameters. Each rig type gets its own PropertyGroup,
so there are no namespace collisions.

This avoids using the ParameterValidator, which is a complex
and evidently not so useful part of Rigify's machinery.

Just as before, when creating a new rig type, it doesn't need to
be added to any sort of global list, it just works.