Initial functioning Generate button
We don't want to use either Rigify's Generator or Rig base classes, so
we take what we need from them, and nothing else.

In this commit, the Generate button no longer fails, although it
only creates an empty armature and assigns it as the target.

There is now also a loop for instantiating rig components
(much simpler than Rigify), although I question if that's even needed.
Perhaps the functions of RigComponents could simply be moved
into PropertyGroups.