Cleanup: Use context.active_object correctly
Previously, I always used context.object.
This is actually the pinned object, when that feature is used.
So, it should only be used in Properties Editor UI.
(This is quite bad design of the PyAPI and that pinning feature)

Also:
- In the 4 UIList drawing codes, use `icon_value` as a more accurate
param name, and put an _ in front of unused params.
- Some other stuff, probably.