Refactor `create_root` param to `ensure_root`.
This now creates a root bone on the metarig, allowing
users to customize the root bone,
and allowing us to make the code a lot nicer, since
we can pass an actual RigComponent and BoneSet
instead of trying to hack them into existence.

That said, the fact that neither of those can exist without a Generator
or RigComponent that owns them respectively,
is pretty major spaghetti.