Colls UX: Better Solo behaviour
I wanted to put this on Ctrl+Click of the eye icon, like in the outliner
but that takes away the ability to click+drag on the eye icons,
which is not worth it.
Still, solo-ing and un-solo-ing now works pretty beautifully, I think,
such that it never messes with the original state of collectio n
visibility, unless the user does as well. I find it even more intuitive
than the Outliner behaviour, actually.