Fix the hotkey system, for real for real.
We do this by not only storing a list of registered addon KeyMapItems in global storage (on a class), but by also first hashing all the inputs of register_hotkey(), and using that hash as a dictionary key.

So, when register_hotkey() is called again with the same params, we catch the matching hash and avoid registering a duplicate hotkey.

Except in some cases, the Reload Scripts operator nukes user KeyMapItems arbitrarily. So it's not enough to check against our own storage, but we then also need to check against Blender's storage, and if Blender nuked out hotkey, we re-create it.

This seems to now support Reload Scripts, as well as toggling CloudRig off and on, as well as spamming the Run Script button on cloudrig.py.

What it does not support, is removing keymap entries on unregister, but that should be quite unnecessary, since they will go away when Blender is restarted.