cloud_spine: Fix issue caused by connected bones.
Previously, we forced ALL bones to use_connected=False.
Removing that caused an issue for the spine rig, where we change the bones' parents, which causes the bones to move.
We now keep bones connected, until their parent is changed. Keeping a bone connected needs to be done explicitly after changing the parent. But this is never done atm.