cloudrig.py: Fix properties drawing in weird order
I didn't expect the fix to be this easy, but it seems that just converting the 'ui_data' nested dictionary to a nested list of tuples works.
In my mind, this should be no different from before, where we were using .keys() to go through the dictionary keys in an arbitrary order. We now use .items() to do the same, and the order is apparently preserved.

Perhaps in Python, .items() is in insertion order, but .keys() is not.
Anyways, while this doesn't feel bulletproof yet, this now allows 'ui_data' to be stored either as a dict or a list of tuples. If we convert the rest of the codebase to feed it a list of tuples, it will definitely work and won't rely on a Python implementation detail.

But first, I want to re-add nesting support, and nuke the legacy panel.