cloudrig.py: Supports more intuitive ui_data
Previously, the ui_data was a nested dictionary where the meaning of each level of nesting was impossible to know without looking at the code that interprets it.
That old format is still supported (although might have some UI cost to draw now).
A new format is also supported, which is essentially a list of tuples, where each element has a type identifier like "header", "property", "properties", "operator", and its corresponding data right next to it. Should be fairly intuitive.

This now also supports drawing any arbitrary property on the rig, not just PoseBone Custom Properties.

Currently, rigs still generate with the old format, and are converted to the new format live, which is not great.

First remedy will be to just do the data conversion once at the end of generation, instead of several times per UI re-draw, that's easy enough.

Second step might be to refactor add_ui_data() to support the new data structure natively, but that's more work.