Properties UI: Operator to edit a slider
This lets us move the slider to a different label or panel, or change which property is being displayed, or change its display name, or all of the above all at once. It's essentially just removing the old property and adding a new one, but it preserves the old one's properties, including its children and operator data.