Symmetrize: Now properly supports Drivers
Used the implementation of Blender's armature_symmetrize_exec
as reference, but applied the value flipping and inverting logic to drivers instead.
This covers cases like a bone's transforms being driven,
IK constraints' pole angle being driven,
Transform constraints, Action constraints.

It considers not only when a value that should be inverted is driven
(in which case the whole expression's value is flipped)
but also when a driver variable is referencing an individual value that should be flipped
(in which case only that variable will be flipped in the expression)

Non-Euler rotations and non-local space are not supported,
there is a warning for the latter.