Chain Rework: Rework spine_squashy
This component type was originally implemented for Charge, ironically enough, a realistic production. And since it was implemented mid-production, it was done quite haphazardly, with a lot of hard-coded logic, assumptions, and limitations. A lot of those should now be lifted, making the squashy spine a nice and simple cartoony spine rig.