Selection Sets are no longer copied from the MetaRig
This was causing an accumulation of selection sets on the target rig. Selection Sets on the generated rig are still preserved as before. The intended use here is that Metarig selection sets are made and used for rigging by the rigger, and selection sets of the generated rig are for animators. The two sets of selection sets are kept completely independent.