Allow overriding generated properties with a Bone Copy component
So for example if you want an IK/FK switch to have a different value than what is normally generated, you can create a bone with the same name as what would be generated (by default "Properties"), create a property on it with the same name as what would be generated, and set whatever default you want on it.

The generation code will simply no longer overwrite existing properties.