Fix #216: Improve detection of installed CloudRig
Previous code assumed the module name is always "CloudRig" since that's the name of the repo, but user can clone the repo with whatever name they please, so needed something more robust.

Tbh it's a blind fix, I didn't verify that this caused the issue or that it fixes it, but I wanna go home and I'd bet 2 nickels that this will do it.