Fix metarig toggle edge case
When there was a non-metarig object named the same as the generated rig just without the RIG- prefix, it would fail to switch to the metarig, and try to switch to that object instead, without even checking if it's an armature.

Now we search for metarigs that reference our rig first, and only if that fails fall back to name-based search, in which case, we now check the object type before trying to switch to it.

Reported by STANN.CO, cheers.