Refactor Constraint Relinking (#219)
Relinking is meant to help with relatively simple cases of wanting to add some extra constraints to the generated rig, but it was unnecessarily limited, and also had a lot of code repetition.

This PR cleans it up, and makes it a bit more powerful. Previously, the target bone to which the constraint was moved was fairly hard-coded for each component type; Constraints added to metarig bones of FK components would be sent to the FK bones, and for Toon chain components, constraints were sent to the STR bones. This is still the default behaviour, but you can now override it by adding any prefix at the start of the constraint name, separated by a dash.

So for example, naming a constraint `RIK-Copy Rotation` on the bone `Foot.L` would send that constraint to the `RIK-Foot.L` bone in the generated rig, whereas previously this was not possible.

Sending constraints with this system to generated bones that don't have an obvious counter-part in the metarig was never possible, and still isn't. For such cases, you have to use the Bone Tweak component, or the post-generation script.

Co-authored-by: Demeter Dzadik <metemer001@gmail.com>
Reviewed-on: https://projects.blender.org/Mets/CloudRig/pulls/219