NEW: cloud_spine_toon component type, for Storm's spine
This new spine component is designed for cartoon characters and comes with some restrictions in order to remain simple; It must consist of exactly 4 bones, and quite a few of its inherited properties are forced to specific values.

Currently the implementation hacks the inherited deform bone set-up of cloud_chain. This hack however seems very promising, and will likely be moved up to cloud_chain as a final solution, just not now.

This component was developed for the Storm character rig, with input from Rik Schutte.

This new implementation might replace spine_squashy in the near future.