Implement get/set_transform on component_type
This means that in future we will be able to change the display names of rig components without metarigs breaking or needing to be versioned.

But for now, metarigs do need to be updated to this new property storage via automatic versioning included in this commit. Only once that's happened on a given metarig, will it be resistant to future renaming of component types. So, let's not rename anything just yet. (I have no plans to anyways.)