WIP: Tests
We desperately need some tests to keep bugs fixed and rigs consistent throughout development.

This is still very early, but it's my top priority for CloudRig.

Currently most of the code is still in the .blend file, and I was trying to implement a pixel comparison based tests relying on viewport playblasting, but the result seems inconsistent, to do either with anti-aliasing or with depth sorting of bone widgets.

I'd also like tests to run automatically periodically (once a day or every push, idk), and have asked Bart for help with that.