Add button to refresh widget list
Sadly this has to be done manually for now, or on a timer, which is sketchy since this process requires checking inside two .blend files on disk, so we don't want to do it too often.

Ideally, this should run when user opens the search menu of any widget selector, but alas, that's not possible.